+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Fix parsing of 'H' 500km squares (Scottish islands)
 - Fix Dms.wrap90(), Dms.wrap180() to work for all -ve degrees
 
+### Added
+
+- LatLonEllipsoidal_Vincenty.intermediatePointTo()
+
 ## [2.2.1] - 2020-04-22
 
 ### Fixed